Structuring the App and using state management to change the state
