+++ index.css
@@ -74,3 +74,9 @@
     display: flex;
     flex-wrap: wrap;
 }
+
+.filter-btn{
+    margin: 10px;
+    padding: 5px;
+    cursor: pointer;
+}
