+++ src/components/Search.js
@@ -1,0 +1,9 @@
+const Search = () => {
+    return (
+        <div className='search'>
+          <h2>Search</h2>
+        </div>
+    );
+}
+
+export default Search;
