+++ src/utilities/constants.js
@@ -1,0 +1,3 @@
+export const res_card_img = "https://media-assets.swiggy.com/swiggy/image/upload/fl_lossy,f_auto,q_auto,w_660/";
+
+export const header_logo = "https://img.freepik.com/free-vector/detailed-chef-logo-template_23-2148987940.jpg?w=740&t=st=1702145429~exp=1702146029~hmac=4e40f0217a8330ed2418695cbb8bae69cf21e6c3a27e96483cb949f1a8a74ec3";
